feat: move tube checking to be after payment and update verify token to be query
