+++ src/constants/paths.ts
@@ -8,4 +8,5 @@ export const PATHS = {
   product: "/kits/:uid",
   root: "/",
   selectedKit: "/selected/:uid",
+  tubes: "/tubes",
 };
