+++ src/store/slices/main-slice.ts
@@ -2,10 +2,12 @@ import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
 
 type initialStateType = {
   currKitUid: string;
+  orderStatus: string;
 };
 
 const initialState: initialStateType = {
   currKitUid: "",
+  orderStatus: "",
 };
 
 export const mainSlice = createSlice({
@@ -15,12 +17,16 @@ export const mainSlice = createSlice({
     setCurrKitUid: (state, { payload }: PayloadAction<string>) => {
       state.currKitUid = payload;
     },
+    setOrderStatus: (state, { payload }: PayloadAction<string>) => {
+      state.orderStatus = payload;
+    },
   },
   selectors: {
     getCurrKitUid: (state) => state.currKitUid,
+    getOrderStatus: (state) => state.orderStatus,
   },
 });
 
-export const { setCurrKitUid } = mainSlice.actions;
-export const { getCurrKitUid } = mainSlice.selectors;
+export const { setCurrKitUid, setOrderStatus } = mainSlice.actions;
+export const { getCurrKitUid, getOrderStatus } = mainSlice.selectors;
 export default mainSlice.reducer;
