+++ src/views/checkout-form/checkout-form.tsx
@@ -1,6 +1,5 @@
 import { Container } from "@chakra-ui/react";
 
-import { CodeCheck } from "#/views/checkout-form/components/code-check";
 import { ConfirmOrder } from "#/views/checkout-form/components/confirm-order";
 import { EmailConfirmation } from "#/views/checkout-form/components/email-confirmation";
 import { OrderDetails } from "#/views/checkout-form/components/order-details";
@@ -9,14 +8,12 @@ import { useAppSelector } from "#store/hooks";
 import { getFormState } from "#store/slices/form-slice";
 
 const formStates = {
-  codeCheck: <CodeCheck />,
   confirmOrder: <ConfirmOrder />,
   emailConfirm: <EmailConfirmation />,
   orderDetails: <OrderDetails />,
 };
 
 const formOrder: formStatesType[] = [
-  "codeCheck",
   "orderDetails",
   "emailConfirm",
   "confirmOrder",
