--- src/views/checkout-form/components/code-check/index.ts
@@ -1,1 +1,0 @@
-export { CodeCheck } from "./code-check";
