+++ src/views/tubes/index.ts
@@ -1,0 +1,1 @@
+export { Tubes } from "./tubes";
